pyproject.toml
[tool.poetry]
name = "noticeboard"
version = "0.1.0"
description = ""
authors = ["Chang-Gyeonghyun <ckh0416@ajou.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
aiomysql = "^0.2.0"
alembic = "^1.13.2"
cryptography = "^43.0.0"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
httpx = "^0.27.0"
redis = "^5.0.8"
aiosqlite = "^0.20.0"
boto3 = "^1.35.5"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
